Add files via upload
Hello everyone. Are you a Linux beginner? If yes then you found the right place for you because here you will be able to find: one of the largest collection of apps/packages that you can be able tyo get onto your computer so that you can not feel like you are struggling.

Our aim is that by the end of the year 2022 we have over 10,000 in our collection that you can be able to download in just 1 minute or less, depending on how many steps there are for the specific program that you might want to download.

Why we have started the Linuxity community is so that maybe one day we can make Linux one of the largest Operating Systems that can be out there and beat Windows so that we can show other people that what people care about is security and not to have the ease of access. But even if you don't have the ease of use at the beginning from the begging, well now you have it, and there IS A LOT of help that you can get from a lot of people that are out there.

At this point in time there are over 44,763,972 users at the point of writing this up (12th March 2022), that might seem like a massive number but that is just a merely 2.2% of the total desktop market share.

With that many users on over 1000+ commercially known distributions and the ones that they might have made on their own is a lot of people for a Operating System that was meant to be for coding only in the 2000s.

In the past 1 year we have come a long way with the new products that have been coming out in 2021/22 like the steam deck and other products that have really pushed the market of semi-smart people to move to Linux.

TILL TODAY!!!

Today we are introducing the easiest way to get programs for Linux.

============================================================================================================

IMPORTANT!!!!!

AT THE MOMENT WE ONLY HAVE FOR ONE BRANCH OF LINUX: UBUNTU. IN THE FUTURE WE WILL TRY TO GET MORE BASE BISTROS TO HAVE THE EASE OF USE WHEN YOU WANT TO GET A LINUX DISTRIBUTION. IN THE FUTURE ON 2023/24 WE ARE LOOKING TO MAYBE SUPPORT OTHER BASE DISTRIBUTIONS LIKE : ARCH,RED HAT AND MANY MORE.

============================================================================================================

You can support us by donating or helping by submitting scripts/codes that people can use to get a program that they might want using the terminal on their preferred flavor/distribution of Linux.

We hope that we can help as many people as possible with the project that we are working on at this point in time.

Hope that you enjoy the work that we put into this tool so that we can help you as a beginner or just for the ease of use with it it on your favorite Linux distribution